rename lib/ansible/galaxy/dependency_resolution/gpg.py -> lib/ansible/galaxy/collection/gpg.py
