+++ lib/ansible/galaxy/collection/gpg.py
